Add enterprise platform phases P0-P6 with data_access hardening and SQL tiers.
Expand query budget, layout policy, materialize services, dependency graph caching, factor lake schema migration, and tier 3-5 SQL pushdown with phase 21-23 platform tests.

+++ factor_engine/cleaned_operators/common/data_cleaning.py
@@ -423,7 +423,7 @@ class IsNaN(SeriesOperator):
     )
 
     def _calculate_series(self, x: pd.DataFrame, **kwargs) -> pd.DataFrame:
-        return x.isna()
+        return x.isna().astype(float)
 
 # aliases: IS_NAN, IS_NULL, is_null
