+++ factor_engine/cleaned_operators/technical/polars_signal.py
@@ -616,7 +616,7 @@ class IsFinitePolars(SeriesOperator):
 
     def _calculate_series(self, x: pl.DataFrame, **kwargs) -> pl.DataFrame:
         cols = _numeric_cols(x)
-        return x.with_columns([pl.col(c).is_finite().cast(pl.Int32).alias(c) for c in cols])
+        return x.with_columns([pl.col(c).is_finite().cast(pl.Float64).alias(c) for c in cols])
 
 
 @register_operator(name="saturate", category="signal", business_category="technical_signal", canonical="saturate", source="factor_dsl_polars")
